RF: Move wtf() out of plugins
Towards a fix for #4519
