+++ datalad/interface/__init__.py
@@ -66,7 +66,7 @@ _group_metadata = (
 _group_misc = (
     'Miscellaneous commands',
     [
-        ('datalad.plugin.wtf', 'WTF'),
+        ('datalad.local.wtf', 'WTF'),
         ('datalad.interface.test', 'Test'),
         ('datalad.interface.ls', 'Ls'),
         ('datalad.interface.clean', 'Clean'),
